requirements.txt
streamlit==1.34.0
openai==1.30.1
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.8.4
